+++ check_workflow/workflow.py
@@ -1,3 +1,4 @@
+import string
 import typing as t
 from collections import defaultdict
 from pathlib import Path
@@ -14,31 +15,43 @@ from check_workflow.gh_api import Release, fetch_releases
 class UsesSpec(t.NamedTuple):  # noqa: D101
     owner: str
     repo: str
-    spec: SpecifierSet
+    spec: SpecifierSet | None
+    sha: str | None
 
     @classmethod
     def from_raw(cls, raw_spec: str) -> t.Self:
         """
         Build a `UsesSpec` instance from the provided workflow dependency specification.
 
-        Dependency specifications are assumed to be of the form `<user>/<repo>@<ver>`, for example:
+        Dependency specifications are assumed to be of the form `<user>/<repo>@<ver>` or a full SHA
+        hash, e.g. :
             * `"actions/setup-python@v6"`
             * `"deadsnakes/action@v3.2.0"`
+            * `"actions/checkout@8f4b7f84864484a7bf31766abe9204da3cbe65b3"`
 
-        The resulting instance's `spec` attribute is built using a compatible release clause (`~=`).
+        If a version specifier is used, the resulting instance's `spec` attribute is built using a
+        compatible release clause (`~=`) and the `sha` attribute will be `None`. Otherwise, `spec`
+        will be `None` and `sha` will be the pinned SHA.
         """
         action, raw_ver = raw_spec.split("@")
         *_, raw_ver = raw_ver.split("/")  # May use a branch, which we don't care about
-        raw_ver = raw_ver.removeprefix("v")  # Some repos may prefix their tags
         owner, repo = action.split("/")
 
-        if len(raw_ver.split(".")) == 1:
-            # Major-only version spec needs special handling, otherwise SpecifierSet will raise
-            spec = SpecifierSet(f"~={raw_ver}.0")
+        # Discriminate between version pin and SHA
+        if len(raw_ver) == 40 and all(c in string.hexdigits for c in raw_ver):
+            spec = None
+            sha = raw_ver
         else:
-            spec = SpecifierSet(f"~={raw_ver}")
+            sha = None
+            raw_ver = raw_ver.removeprefix("v")  # Some repos may prefix their tags
 
-        return cls(owner=owner, repo=repo, spec=spec)
+            if len(raw_ver.split(".")) == 1:
+                # Major-only version spec needs special handling, otherwise SpecifierSet will raise
+                spec = SpecifierSet(f"~={raw_ver}.0")
+            else:
+                spec = SpecifierSet(f"~={raw_ver}")
+
+        return cls(owner=owner, repo=repo, spec=spec, sha=sha)
 
 
 class JobDependency(t.NamedTuple):  # noqa: D101
@@ -48,7 +61,12 @@ class JobDependency(t.NamedTuple):  # noqa: D101
 
 
 def extract_workflow_dependencies(raw_workflow: str) -> list[JobDependency]:
-    """Extract job dependencies from the provided raw workflow YAML."""
+    """
+    Extract job dependencies from the provided raw workflow YAML.
+
+    NOTE: Only versioned actions are considered. Other specifications, such as local or docker
+    actions, are skipped.
+    """
     loaded = yaml.safe_load(raw_workflow)
 
     extracted_dependencies = []
@@ -56,14 +74,21 @@ def extract_workflow_dependencies(raw_workflow: str) -> list[JobDependency]:
         for step in job_params["steps"]:
             uses = step.get("uses", None)
 
-            if uses:
-                extracted_dependencies.append(
-                    JobDependency(
-                        job=job,
-                        step_name=step.get("name", None),
-                        uses=UsesSpec.from_raw(uses),
-                    )
+            if uses is None:
+                continue
+
+            if uses.startswith("docker"):
+                continue
+            elif uses.startswith("./") or uses.startswith("../"):
+                continue
+
+            extracted_dependencies.append(
+                JobDependency(
+                    job=job,
+                    step_name=step.get("name", None),
+                    uses=UsesSpec.from_raw(uses),
                 )
+            )
 
     return extracted_dependencies
 
@@ -95,13 +120,26 @@ async def report_outdated(
             else:
                 latest = seen_releases[cache_key]
 
-            if latest.ver not in dep.uses.spec:
+            # Switch behavior based on whether we've pinned a version vs. SHA
+            # If sha is None then spec is defined & vice-versa; since this is the only place this
+            # comparison happens we can go with this assumption vs. adding more narrowing logic
+            is_outdated = False
+            if dep.uses.sha is None:
+                if latest.ver not in dep.uses.spec:  # type: ignore[operator]
+                    is_outdated = True
+            else:
+                if latest.tag_hash != dep.uses.sha:
+                    is_outdated = True
+
+            if is_outdated:
                 outdated[wf_name].append(OutdatedDep(spec=dep, latest=latest))
 
     return outdated
 
 
-def format_outdated(outdated: dict[str, list[OutdatedDep]], markdown: bool = False) -> str:
+def format_outdated(
+    outdated: dict[str, list[OutdatedDep]], markdown: bool = False
+) -> str:  # pragma: no cover
     """
     Prettyprint the provided outdated dependencies into a table.
 
@@ -125,13 +163,22 @@ def format_outdated(outdated: dict[str, list[OutdatedDep]], markdown: bool = Fal
 
         for dep in deps:
             action_string = f"{dep.spec.uses.owner}/{dep.spec.uses.repo}"
+
+            # If spec is None then SHA is defined & vice-versa
+            if dep.spec.uses.spec is not None:
+                uses_spec = dep.spec.uses.spec
+                latest_spec = dep.latest.ver
+            else:
+                uses_spec = dep.spec.uses.sha[:7]  # type: ignore[index,assignment]
+                latest_spec = dep.latest.tag_hash[:7]  # type: ignore[assignment]
+
             table.add_row(
                 [
                     dep.spec.job,
                     dep.spec.step_name,
                     action_string,
-                    dep.spec.uses.spec,
-                    dep.latest.ver,
+                    uses_spec,
+                    latest_spec,
                 ]
             )